+++ src/components/Template/TemplateContent.tsx
@@ -1,10 +1,10 @@
-import { Box } from '@chakra-ui/react';
 import styled from '@emotion/styled';
 import { useState } from 'react';
 
 import EmptyQuestion from '@/assets/EmptyQuestion.svg';
 
 import TemplateFooter from './TemplateFooter';
+import TemplateQuestionSelections from './TemplateQuestionSelections';
 import TemplateSelections from './TemplateSelections';
 import TemplateSubHeader from './TemplateSubHeader';
 
@@ -27,43 +27,7 @@ export default function TemplateContent({
       ) : (
         <>
           <TemplateSubHeader />
-          <QeustionsContainer>
-            <Box
-              as="button"
-              padding={1}
-              px={'14px'}
-              fontSize={'14px'}
-              border={'1px'}
-              borderRadius={'20px 20px 20px 20px'}
-              borderColor={'#ccd0d5'}
-              _hover={{
-                bg: '#6691FF',
-                borderColor: '#6691FF',
-                color: '#FFFFFF',
-              }}
-              marginRight={2}
-              onClick={() => setQuestions(['temporary'])}
-            >
-              + 기본 문항
-            </Box>
-            <Box
-              as="button"
-              padding={1}
-              border={'1px'}
-              px={'14px'}
-              fontSize={'14px'}
-              borderRadius={'20px 20px 20px 20px'}
-              bg={'none'}
-              borderColor={'#ccd0d5'}
-              _hover={{
-                bg: '#1FB881',
-                borderColor: '#1FB881',
-                color: '#FFFFFF',
-              }}
-            >
-              + 전문 문항
-            </Box>
-          </QeustionsContainer>
+          <TemplateQuestionSelections setQuestions={setQuestions} />
           <ContentContainer>
             {questions.length === 0 && (
               <EmptyQuestionContainer>
@@ -82,12 +46,6 @@ export default function TemplateContent({
     </div>
   );
 }
-
-const QeustionsContainer = styled('div')`
-  display: flex;
-  justify-content: flex-end;
-  margin: 0 2rem 1rem 2rem;
-`;
 
 const ContentContainer = styled('div')`
   display: flex;
