+++ src/components/Template/index.tsx
@@ -1,0 +1,22 @@
+import TemplateContent from './TemplateContent';
+import TemplateTitle from './TemplateTitle';
+
+type TemplateProps = {
+  selectedTemplateTitle: string;
+  setSelectedTemplateTitle: React.Dispatch<React.SetStateAction<string>>;
+};
+
+export default function Template({
+  selectedTemplateTitle,
+  setSelectedTemplateTitle,
+}: TemplateProps) {
+  return (
+    <>
+      <TemplateTitle title={selectedTemplateTitle} />
+      <TemplateContent
+        selectedTemplateTitle={selectedTemplateTitle}
+        setSelectedTemplateTitle={setSelectedTemplateTitle}
+      />
+    </>
+  );
+}
