+++ src/pages/RecordManagementPage.tsx
@@ -1,10 +1,19 @@
 import { Button } from '@chakra-ui/react';
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 
 import Modal from '@/components/common/Modal';
+import Template from '@/components/Template';
 
 export default function RecordManagementPage() {
   const [isOpen, setIsOpen] = useState(false);
+
+  const [selectedTemplateTitle, setSelectedTemplateTitle] = useState('');
+
+  useEffect(() => {
+    if (!isOpen) {
+      setSelectedTemplateTitle('');
+    }
+  }, [isOpen]);
 
   return (
     <div>
@@ -20,7 +29,18 @@ export default function RecordManagementPage() {
       >
         + 템플릿 추가
       </Button>
-      {isOpen && <Modal width={700} height={400} setIsOpen={setIsOpen} />}
+      {isOpen && (
+        <Modal
+          width={selectedTemplateTitle.length === 0 ? 700 : undefined}
+          height={selectedTemplateTitle.length === 0 ? 400 : undefined}
+          setIsOpen={setIsOpen}
+        >
+          <Template
+            selectedTemplateTitle={selectedTemplateTitle}
+            setSelectedTemplateTitle={setSelectedTemplateTitle}
+          />
+        </Modal>
+      )}
     </div>
   );
 }
